+++ varnish-sys/src/txt.rs
@@ -45,19 +45,18 @@ impl txt {
         }
     }
 
-    /// Convert the `txt` struct to a `&str`.  Will panic if the string is not valid UTF-8.
-    #[expect(clippy::wrong_self_convention)] // TODO: drop Copy trait for txt?
-    pub fn to_str<'a>(&self) -> Option<&'a str> {
-        self.to_slice().map(|s| from_utf8(s).unwrap())
-    }
-
     /// Parse the `txt` struct as a header, returning a tuple with the key and value,
     /// trimming the value of leading whitespace.
-    pub fn parse_header<'a>(&self) -> Option<(&'a str, &'a str)> {
+    pub fn parse_header<'a>(&self) -> Option<(&'a str, &'a [u8])> {
         // We expect varnishd to always given us a string with a ':' in it
         // If it's not the case, blow up as it might be a sign of a bigger problem.
-        let (key, value) = self.to_str()?.split_once(':').unwrap();
+        let slice = self.to_slice()?;
+        let index = slice.iter().position(|c| *c == b':').expect("headers should always have a :");
+        let (key_slice, value_slice) = slice.split_at(index);
+
+        let key = from_utf8(key_slice).expect("header names must be UTF-8");
         // FIXME: Consider `.trim_ascii_start()` if unicode is not a concern
-        Some((key, value.trim_start()))
+        let value = value_slice[1..].trim_ascii_start();
+        Some((key, value))
     }
 }