+++ varnish/tests/vmod_test/src/lib.rs
@@ -140,6 +140,14 @@ mod rustest {
         }
     }
 
+    pub fn merge_all_names(ctx: &Ctx) -> String {
+        let mut s = String::new();
+        for (k, _) in ctx.http_req.as_ref().expect("merge_all_names is only available in a client context") {
+            s += k;
+        }
+        s
+    }
+
     #[event]
     pub fn event(event: Event, vfp: &mut FetchFilters) {
         if let Event::Load = event {