+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - The returned buffers support `Write` trait, and can be inspected/modified what has been written so far
   - The buffer does not allow any access to "dirty" (unset) portion of the buffer
   - The buffer must be finalized with `finish()`, which returns `VCL_STRING`, `VCL_BLOB`, or `&[T]` depending on the builder used
+- Header values are now `&[u8]` instead of `&str`
 - Remove `vsc` feature - all of its functionality is now available without any feature flags
 - Rename `Stat` &rarr; `Metrics`, `Stats` &rarr; `MetricsReader`, `StatsBuilder` &rarr; `MetricsReaderBuilder`, and `Format` &rarr; `MetricsFormat`
 - `MetricsReaderBuilder::patience` now returns `Self`